MedfordCharities/requirements.txt
flask
google-cloud-datastore
requests
